# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Objective Financial API",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/api/account": {
            "get": {
                "tags": [
                    "accounts"
                ],
                "summary": "Get all accounts",
                "operationId": "9c3e3a19f5dee65a542defa8ca8dbb26",
                "parameters": [
                    {
                        "name": "account_number",
                        "in": "query",
                        "description": "Account number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "account_number": {
                                                "type": "integer"
                                            },
                                            "balance": {
                                                "type": "number",
                                                "format": "float"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "accounts"
                ],
                "summary": "Create a new account",
                "operationId": "648f9a087a595875582d2780ccc5e4d4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "account_number": {
                                        "type": "integer"
                                    },
                                    "balance": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "account_number": {
                                            "type": "integer"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/account/{id}": {
            "get": {
                "tags": [
                    "accounts"
                ],
                "summary": "Get account by id",
                "operationId": "cd9f0183c5cd7bac67286f466d1597bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "account_number": {
                                            "type": "integer"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "tags": [
                    "accounts"
                ],
                "summary": "Update account by id",
                "operationId": "8ef1d01b5c524f0ee8a3554969745fe5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "account_number": {
                                        "type": "integer"
                                    },
                                    "balance": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "account_number": {
                                            "type": "integer"
                                        },
                                        "balance": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "accounts"
                ],
                "summary": "Delete account by id",
                "operationId": "04193849c33df843473d0945b009a9f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "accounts",
            "description": "accounts"
        }
    ]
}